Move ThisApi to shared context
